update respondents and child controller tests, fix bugs in RepresentableLegalCounselUpdater
