+++ service/src/integrationTest/java/uk/gov/hmcts/reform/fpl/controllers/AbstractTest.java
@@ -20,7 +20,6 @@ import java.time.LocalDateTime;
 import java.util.Map;
 
 import static org.mockito.BDDMockito.given;
-import static uk.gov.hmcts.reform.fpl.handlers.NotificationEventHandlerTestData.AUTH_TOKEN;
 
 @ActiveProfiles("integration-test")
 public abstract class AbstractTest {
@@ -76,8 +75,9 @@ public abstract class AbstractTest {
     }
 
     protected void givenSystemUser() {
-        given(idamClient.getAccessToken(userConfig.getUserName(), userConfig.getPassword())).willReturn(AUTH_TOKEN);
-        given(idamClient.getUserInfo(AUTH_TOKEN)).willReturn(UserInfo.builder()
+        given(idamClient.getAccessToken(userConfig.getUserName(), userConfig.getPassword()))
+            .willReturn(USER_AUTH_TOKEN);
+        given(idamClient.getUserInfo(USER_AUTH_TOKEN)).willReturn(UserInfo.builder()
             .uid(SYS_USER_ID)
             .build());
     }
