+++ service/src/integrationTest/java/uk/gov/hmcts/reform/fpl/controllers/ChildControllerSubmittedTest.java
@@ -8,6 +8,9 @@ import org.springframework.boot.test.autoconfigure.OverrideAutoConfiguration;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.boot.test.mock.mockito.SpyBean;
+import uk.gov.hmcts.reform.ccd.client.CaseAccessDataStoreApi;
+import uk.gov.hmcts.reform.ccd.model.CaseAssignedUserRoleWithOrganisation;
+import uk.gov.hmcts.reform.ccd.model.CaseAssignedUserRolesRequest;
 import uk.gov.hmcts.reform.ccd.model.ChangeOrganisationRequest;
 import uk.gov.hmcts.reform.ccd.model.Organisation;
 import uk.gov.hmcts.reform.fpl.enums.State;
@@ -15,8 +18,10 @@ import uk.gov.hmcts.reform.fpl.events.AfterSubmissionCaseDataUpdated;
 import uk.gov.hmcts.reform.fpl.model.CaseData;
 import uk.gov.hmcts.reform.fpl.model.Child;
 import uk.gov.hmcts.reform.fpl.model.ChildParty;
+import uk.gov.hmcts.reform.fpl.model.LegalCounsellor;
 import uk.gov.hmcts.reform.fpl.model.RespondentSolicitor;
 import uk.gov.hmcts.reform.fpl.model.UnregisteredOrganisation;
+import uk.gov.hmcts.reform.fpl.model.common.Element;
 import uk.gov.hmcts.reform.fpl.model.common.dynamic.DynamicList;
 import uk.gov.hmcts.reform.fpl.model.common.dynamic.DynamicListElement;
 import uk.gov.hmcts.reform.fpl.model.event.ChildrenEventData;
@@ -24,7 +29,10 @@ import uk.gov.hmcts.reform.fpl.service.EventService;
 import uk.gov.hmcts.reform.fpl.service.FeatureToggleService;
 import uk.gov.hmcts.reform.fpl.service.NoticeOfChangeService;
 import uk.gov.hmcts.reform.fpl.service.ccd.CoreCaseDataService;
+import uk.gov.hmcts.reform.rd.client.OrganisationApi;
+import uk.gov.hmcts.reform.rd.model.OrganisationUser;
 import uk.gov.service.notify.NotificationClient;
+import uk.gov.service.notify.NotificationClientException;
 
 import java.util.HashMap;
 import java.util.List;
@@ -42,6 +50,7 @@ import static uk.gov.hmcts.reform.fpl.CaseDefinitionConstants.JURISDICTION;
 import static uk.gov.hmcts.reform.fpl.Constants.LOCAL_AUTHORITY_1_CODE;
 import static uk.gov.hmcts.reform.fpl.Constants.LOCAL_AUTHORITY_1_COURT_NAME;
 import static uk.gov.hmcts.reform.fpl.Constants.LOCAL_AUTHORITY_1_NAME;
+import static uk.gov.hmcts.reform.fpl.NotifyTemplates.LEGAL_COUNSELLOR_REMOVED_EMAIL_TEMPLATE;
 import static uk.gov.hmcts.reform.fpl.NotifyTemplates.REGISTERED_RESPONDENT_SOLICITOR_TEMPLATE;
 import static uk.gov.hmcts.reform.fpl.NotifyTemplates.UNREGISTERED_RESPONDENT_SOLICITOR_TEMPLATE;
 import static uk.gov.hmcts.reform.fpl.enums.SolicitorRole.CHILDSOLICITORA;
@@ -91,6 +100,10 @@ class ChildControllerSubmittedTest extends AbstractCallbackTest {
     private FeatureToggleService toggleService;
     @MockBean
     private NotificationClient notificationClient;
+    @MockBean
+    private OrganisationApi orgApi;
+    @MockBean
+    private CaseAccessDataStoreApi accessApi;
 
     ChildControllerSubmittedTest() {
         super("enter-children");
@@ -98,6 +111,11 @@ class ChildControllerSubmittedTest extends AbstractCallbackTest {
 
     @BeforeEach
     void setUp() {
+        givenFplService();
+        givenSystemUser();
+        when(orgApi.findOrganisation(USER_AUTH_TOKEN, SERVICE_AUTH_TOKEN, ORGANISATION_ID))
+            .thenReturn(uk.gov.hmcts.reform.rd.model.Organisation.builder().name(ORGANISATION_NAME).build());
+
         when(toggleService.isChildRepresentativeSolicitorEnabled()).thenReturn(true);
     }
 
@@ -331,6 +349,79 @@ class ChildControllerSubmittedTest extends AbstractCallbackTest {
                 ),
                 "localhost/" + CASE_ID
             )
+        );
+    }
+
+    @Test
+    void shouldRevokeAccessAndSendNotificationsWhenLegalCounselRemoved() throws NotificationClientException {
+        final String legalCounsellorId = "some id";
+        final String legalCounsellorEmail = "some email";
+        final List<Element<LegalCounsellor>> legalCounsellors = wrapElements(
+            LegalCounsellor.builder().firstName("first").lastName("last").email(legalCounsellorEmail).build()
+        );
+
+        when(orgApi.findUserByEmail(USER_AUTH_TOKEN, SERVICE_AUTH_TOKEN, legalCounsellorEmail))
+            .thenReturn(OrganisationUser.builder().userIdentifier(legalCounsellorId).build());
+
+        CaseData caseDataBefore = CaseData.builder()
+            .state(NON_RESTRICTED_STATE)
+            .id(CASE_ID)
+            .caseName(CASE_NAME)
+            .caseLocalAuthority(LOCAL_AUTHORITY_1_CODE)
+            .children1(wrapElements(
+                Child.builder()
+                    .party(
+                        ChildParty.builder()
+                            .firstName(CHILD_NAME_1)
+                            .lastName(CHILD_SURNAME_1)
+                            .dateOfBirth(dateNow())
+                            .build()
+                    )
+                    .solicitor(REGISTERED_REPRESENTATIVE)
+                    .legalCounsellors(legalCounsellors)
+                    .build()
+            ))
+            .build();
+
+        CaseData caseData = caseDataBefore.toBuilder()
+            .children1(wrapElements(
+                Child.builder()
+                    .party(
+                        ChildParty.builder()
+                            .firstName(CHILD_NAME_1)
+                            .lastName(CHILD_SURNAME_1)
+                            .dateOfBirth(dateNow())
+                            .build()
+                    )
+                    .solicitor(UNREGISTERED_REPRESENTATIVE)
+                    .build()
+            ))
+            .build();
+
+        postSubmittedEvent(toCallBackRequest(caseData, caseDataBefore));
+
+        CaseAssignedUserRolesRequest revokeRequestPayload = CaseAssignedUserRolesRequest.builder()
+            .caseAssignedUserRoles(List.of(
+                CaseAssignedUserRoleWithOrganisation.builder()
+                    .userId(legalCounsellorId)
+                    .caseRole("[BARRISTER]")
+                    .caseDataId(CASE_ID.toString())
+                    .build()
+            ))
+            .build();
+
+        verify(accessApi).removeCaseUserRoles(USER_AUTH_TOKEN, SERVICE_AUTH_TOKEN, revokeRequestPayload);
+
+        Map<String, Object> notifyData = Map.of(
+            "caseName", CASE_NAME,
+            "ccdNumber", "1234-5678-9012-3456",
+            "childLastName", CHILD_SURNAME_1,
+            "clientFullName", ORGANISATION_NAME,
+            "salutation", "Dear first last"
+        );
+
+        verify(notificationClient).sendEmail(
+            LEGAL_COUNSELLOR_REMOVED_EMAIL_TEMPLATE, legalCounsellorEmail, notifyData, "localhost/" + CASE_ID
         );
     }
 }
