+++ service/src/integrationTest/java/uk/gov/hmcts/reform/fpl/controllers/RespondentControllerAboutToStartTest.java
@@ -1,0 +1,31 @@
+package uk.gov.hmcts.reform.fpl.controllers;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.autoconfigure.OverrideAutoConfiguration;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
+import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
+
+import java.util.Map;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+@WebMvcTest(RespondentController.class)
+@OverrideAutoConfiguration(enabled = true)
+class RespondentControllerAboutToStartTest extends AbstractCallbackTest {
+
+    RespondentControllerAboutToStartTest() {
+        super("enter-respondents");
+    }
+
+    @Test
+    void aboutToStartShouldPrePopulateRespondent() {
+        CaseDetails caseDetails = CaseDetails.builder()
+            .data(Map.of("data", "some data"))
+            .build();
+
+        AboutToStartOrSubmitCallbackResponse callbackResponse = postAboutToStartEvent(caseDetails);
+
+        assertThat(callbackResponse.getData()).containsKey("respondents1");
+    }
+}
