+++ service/src/main/java/uk/gov/hmcts/reform/fpl/service/legalcounsel/RepresentableLegalCounselUpdater.java
@@ -21,6 +21,7 @@ import uk.gov.hmcts.reform.fpl.service.UserService;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -41,65 +42,34 @@ public class RepresentableLegalCounselUpdater {
     private final OrganisationService orgService;
     private final UserService user;
 
-    public List<LegalCounsellorRemoved> buildEventsForAccessRemoval(CaseData caseData, CaseData caseDataBefore,
-                                                                    final SolicitorRole.Representing representing) {
-        // need to find the case role that has been updated
+    public Set<LegalCounsellorRemoved> buildEventsForAccessRemoval(CaseData caseData, CaseData caseDataBefore,
+                                                                          final SolicitorRole.Representing representing) {
+        // need to find the representables that have been updated
         List<WithSolicitor> current = unwrapElements(representing.getTarget().apply(caseData));
         List<WithSolicitor> old = unwrapElements(representing.getTarget().apply(caseDataBefore));
-
-        WithSolicitor changed = null;
-
-        for (int i = 0; i < current.size(); i++) {
-            WithSolicitor representable = current.get(i);
-            if (!Objects.equals(representable.getSolicitor(), old.get(i).getSolicitor())) {
-                changed = representable;
-                break;
-            }
-        }
-
-        if (null == changed) {
-            log.debug("No change in solicitor detected for {}", representing.name());
-            return List.of();
-        }
-
-        String orgName = getOrgId(changed)
-            .map(orgService::findOrganisation)
-            .filter(Optional::isPresent)
-            .map(Optional::get)
-            .map(uk.gov.hmcts.reform.rd.model.Organisation::getName)
-            .orElseGet(() -> {
-                // in the scenario that a user has removed the current solicitor (possible in respondent flow)
-                // Refer to all HMCTS staff as just HMCTS to cover the scenario that it is an admin performing the
-                // action (possibly on behalf of a judge)
-                // otherwise it must be a solicitor, and they must be associated to organisation
-                if (user.isHmctsUser()) {
-                    return "HMCTS";
-                }
-                return orgService.findOrganisation()
-                    .map(uk.gov.hmcts.reform.rd.model.Organisation::getName)
-                    .orElseThrow(() -> new IllegalStateException(
-                        "Solicitor was changed to null, the user was not HMCTS, the user does not have a organisation"
-                        + " associated to them"
-                    ));
-            });
 
         Set<LegalCounsellor> allCurrentLegalCounsellors = Arrays.stream(SolicitorRole.Representing.values())
             .flatMap(r -> unwrapElements(r.getTarget().apply(caseData)).stream())
             .flatMap(s -> unwrapElements(s.getLegalCounsellors()).stream())
             .collect(Collectors.toSet());
 
-        Set<LegalCounsellor> allOldLegalCounsellors = Arrays.stream(SolicitorRole.Representing.values())
-            .flatMap(r -> unwrapElements(r.getTarget().apply(caseDataBefore)).stream())
-            .flatMap(s -> unwrapElements(s.getLegalCounsellors()).stream())
-            .collect(Collectors.toSet());
+        Set<LegalCounsellorRemoved> events = new HashSet<>();
 
-        return allOldLegalCounsellors.stream()
-            .filter(not(allCurrentLegalCounsellors::contains))
-            .map(counsellor -> orgService.findUserByEmail(counsellor.getEmail()).map(id -> Pair.of(id, counsellor)))
-            .filter(Optional::isPresent)
-            .map(Optional::get)
-            .map(pair -> new LegalCounsellorRemoved(caseData, orgName, pair))
-            .collect(Collectors.toList());
+        for (int i = 0; i < old.size(); i++) {
+            WithSolicitor currentRepresentable = current.get(i);
+            WithSolicitor oldRepresentable = old.get(i);
+            if (!Objects.equals(currentRepresentable.getSolicitor(), oldRepresentable.getSolicitor())) {
+                events.addAll(unwrapElements(oldRepresentable.getLegalCounsellors()).stream()
+                    .filter(not(allCurrentLegalCounsellors::contains))
+                    .map(this::mapLegalCounsellor)
+                    .filter(Optional::isPresent)
+                    .map(Optional::get)
+                    .map(pair -> new LegalCounsellorRemoved(caseData, getOrgName(currentRepresentable), pair))
+                    .collect(Collectors.toSet()));
+            }
+        }
+
+        return events;
     }
 
     public Map<String, Object> updateLegalCounsel(CaseData caseData, CaseData caseDataBefore) {
@@ -183,5 +153,34 @@ public class RepresentableLegalCounselUpdater {
             .map(WithSolicitor::getSolicitor)
             .map(RespondentSolicitor::getOrganisation)
             .map(Organisation::getOrganisationID);
+    }
+
+    private Optional<Pair<String, LegalCounsellor>> mapLegalCounsellor(LegalCounsellor counsellor) {
+        return orgService.findUserByEmail(counsellor.getEmail()).map(id -> Pair.of(id, counsellor));
+    }
+
+    private String getOrgName(WithSolicitor changed) {
+        return getOrgId(changed)
+            .flatMap(orgService::findOrganisation)
+            .map(uk.gov.hmcts.reform.rd.model.Organisation::getName)
+            .orElseGet(() -> {
+                RespondentSolicitor solicitor = changed.getSolicitor();
+                if (null != solicitor && null != solicitor.getUnregisteredOrganisation()) {
+                    return solicitor.getUnregisteredOrganisation().getName();
+                }
+                // in the scenario that a user has removed the current solicitor (possible in respondent flow)
+                // Refer to all HMCTS staff as just HMCTS to cover the scenario that it is an admin performing the
+                // action (possibly on behalf of a judge)
+                // otherwise it must be a solicitor, and they must be associated to organisation
+                if (user.isHmctsUser()) {
+                    return "HMCTS";
+                }
+                return orgService.findOrganisation()
+                    .map(uk.gov.hmcts.reform.rd.model.Organisation::getName)
+                    .orElseThrow(() -> new IllegalStateException(
+                        "Solicitor was changed to null, the user was not HMCTS, the user does not have a organisation"
+                        + " associated to them"
+                    ));
+            });
     }
 }
